backend/requirements.txt
django==4.0.2
djangorestframework